Updated openstack/openstack Project: openstack/taskflow  b18798520aabe320b556e45cd86403692d424408
Use a single shared queue for an executors lifecycle

Instead of having many queues (one per task) just have a shared
one and use sender identifiers to know which originating task
to proxy back to.

This scales better and avoids the needless polling of many
queues for a potential message to emit (and can now instead
just poll one queue); when now we can just poll a single
queue for any messages (this does though remove the ability to
throttle messages from a given sender).

Change-Id: I3566a5ab20ad15a80a4c6969f48b076ddde1d7ac