Updated openstack/openstack Project: openstack/nova  5a6b8ff00f80eb61fb74975dd7e5749007e88edc
Make neutronapi get networks operations return objects

This makes the neutronapi get() and get_all() methods return objects
instead of bare dicts. Note that previously we were just returning
(basically) the raw result of the neutron operation, which doesn't
actually map to what nova-network will return. This helps to
unify some of that by constructing the object to look like we'd
expect it to from what we get from Neutron (minus all the bits we
don't have of course).

Related to blueprint kilo-objects

Change-Id: Ied70ab493982896cd974bb5a98ca92d95352e80f