Updated openstack/openstack Project: openstack/tempest  bdf9bc2ad4164c55653140eb7d959365bfd734c6
Ensure subnet for port security group tests

This change simply adds a subnet to the networks in tests
verifying creation or update of a port without a security
group. This is being done because there are slight differences
across neutron plugins logic for managing association between
ports and security groups. For instance, the VMware NSX plugin
which also implements the "port security" extension, requires
an IP address on a port in order to associate it with the
default security group.

Adding a subnet to the network used in the test ensures the
test exercises a use case which is common to all plugins.

This patch simply adds the subnet to the network without
changing the remaining logic of the modified tests.

Change-Id: Ic1f725f5a307ac74acfb13b831db09025025ea2c