Updated openstack/openstack Project: openstack/taskflow  d9fd9eed6106f0b840e264482fa2bb64218a5ec1
Allow specifying the engine 'executor' as a string

To enable a parallel process executor to be used without
having to pass in a futures executor allow for the executor
option to be a string 'processes' (or similar) that will cause
the default parallel process executor to be used automatically.

Also allow for a 'threads' string that ensure a parallel thread
executor is used to match the ability to uses processes.

This also adjusts the WBE engine to have a similar executor fetching
function (which in the WBE case now validates a provided executor to
be of the desired type).

Change-Id: I54a82584c32c697922507b4f6e01ea7b8acc73c6