Updated openstack/openstack Project: openstack-dev/devstack  5c0a3cfb71ef8c8faca92c19e4eee4968eb309e3
Install Juno RDO repos for RHEL7

RHEL7 does not include Open vSwitch in the default repos, but it
is available via the RDO repo.  This patch automatically
configures the RDO repo for RHEL7.

Move this into the existing rhel6/rhel7 section.  We update RHEl6 to
the latest icehouse release, but it is not supported with Juno

Closes-Bug: #1402390
Change-Id: I4707cf68e39d9e900ec6c01331d7e124c8c4f6c4