Updated openstack/openstack Project: openstack/keystone  52738c6f5ab4383b3c9fbb170a3aeded7a8b2786
switch from sample_config.sh to oslo-config-generator

This switches keystone to use oslo-config-generator, where the contents
of our sample configuration file are configured using a configuration
file in config-generators/, because the only way to have tox create
files is to have configuration files for your configuration files
(seriously, it chokes when redirects are included in the command for a
test environment).

This also takes advantage of oslo-config-generator's ability to set the
line wrap width, which is done at 79 chars here to match the rest of our
code.

Also update the actual sample file under etc/

Change-Id: I9f9a157c7802c56a2c7484f6f4adec4aeb4071b8
Closes-Bug: 1324610