Updated openstack/openstack Project: openstack/horizon  d7a13e90ddf4e07cc13ecfc31b89ce67e60afc85
Fix wrong calling of selenium driver function

Selenium docstrings define the usage of this function to be called as
driver.current_url but integration tests call it in wrong way as
driver.current_url() which throws an error if you try to use it that way.
Read the bug for detailed info on this and its source.

Change-Id: I9b065ffb495b637dfe4582edf4aa216a833d5058
Partially implements blueprint: selenium-integration-testing
Closes-Bug: #1404379