Updated openstack/openstack Project: openstack/oslo-incubator  103ebd52480f2794f99242e6ad2a9d2ae9dc1024
Make sure run_cross_tests.sh installs all files

Rebuild the egg_info directory for the package to ensure that the list
of source files is up to date before pip tries to install the package.
This ensures that tests for libs where we are renaming the python
package install all of the files in their new locations.

Change-Id: If5fb35ea7e2bafa5586850b9a42beff60be3dbd9