Updated openstack/openstack Project: openstack/heat  cea0a0f7bb498b0fd20432467c42eb176e01c11b
Gracefully fail to delete nonempty S3 Bucket

The AWS::S3::Bucket resource is susceptible to bug #1406263
as it is backed by the same Swift container - on deleting non empty
container it produces incomprehensible truncated message.

This patch uses the same logic as it is now in OS::Swift::Container
resource, making the reason of failure clear. The error message is made
to resemble the error produced by this resource on real AWS
in the same usage scenario.

Change-Id: Ib67c55f010613c8edb913bb264d3cc4a7315d847
Related-Bug: #1406263