Updated openstack/openstack Project: openstack/nova  3e2e6968b098c8e0722f88cc330ff0c19177266a
hardware: fix numa topology from image meta data

The behavior of the method numa_get_constraints is to get a dict of
image properties to determine whether a NUMA topology has been defined
in the image.
We currently pass to this method a dict of image information or sometimes only
a dict of properties of the image.

We should always pass the dict of image information

Change-Id: If023bfa766f67288699c17f5b5daabdc24d03ea2
Closes-Bug: #1404839