Updated openstack/openstack Project: openstack/trove  06ae4c5358fd544089f851f635ff3090f1ffabb6
Config Group Load Fails If DS Version Inactive

trove-manage db_load_datastore_config_parameters will fail if
the datastore_version on which it's being run is marked as
inactive (aka active is set to 0).

this is not correct, because it's completely reasonable to want
to update/refresh a datastore_version's configuration-group's
parameters even if it's inactive.

examples:
- you're working on rolling out a new datastore_version, so
  you have it marked as active=0 until the go-live date.
- you've rolled out a new datastore_version and want to
  mark the old version as inactive. instances built prior
  to the inactivation are still on the old version, and
  can still utilize their still-assigned configuration-groups,
  hence the parameters should be updateable by the
  administrator.

Change-Id: I7682fe240c5737438b084d813cf7ce80b4ecca9b
Closes-Bug: #1379563