Updated openstack/openstack Project: openstack/oslo.log  651ee37399b1782ec841c93aca3b87d0272a0761
Implement resource to logging extra keywords

Cleaning up log messages and thought it
would be useful to make it "easier" to log the
volume/snapshot/backup/image id's more easily.

Unlike the Nova Instance work that was submitted this
key will work with any db ref object that has a "name"
member.  Most OpenStack objects have an internal name
that is built using the UUID of the resource object.
Resource objects in Cinder for example use:
  "<resourceName-uuid>"

For those that don't have the name member, we also
give the ability to set resource as a dict, with
keys 'type' and 'id' that will form the same string.

With this change, LOG calls in projects can be sent like:
  LOG.info(_LI('begin some operation'), resource=dbref)
or
  LOG.info(_LI('begin some operation'), resource=
           mydict{'type': 'volume',
                  'id': '40e0f194-0ccb-422f-a543-c5f9022acb6f')

The result will be the resource identifier prepended to the msg:

  INFO cinder.volume.manager [-]\
  [volume-40e0f194-0ccb-422f-a543-c5f9022acb6f] \
  succesfully deleted volume

This will be in addition to work being done to put the same info in context,
then we'll automatically pick this info up from context if it's there but still
allow it to be overridden with the extra keys in this change.

Change-Id: I67b31f666172451fd6b061520b0f171ec8e7387b
Partially-Implements: blueprint app-agnostic-logging-parameters