Updated openstack/openstack Project: openstack/nova  b19c0519d729bcb6ff68ee6594e8803acfcf6ff3
Support both list and dict for pci_passthrough_whitelist

In Icehouse, pci_passthrough_whitelist is a json docstring that
encodes a list. In Juno, it is a json docstring that encodes
a dict. This patch adds the list support back to
pci_passthrough_whitelist, and both list and dict are now supported.

Change-Id: I523cfb756a09c75e4f60015adadb3a1403298cd3
Closes-Bug: 1383345