Updated openstack/openstack Project: openstack/cinder  894f20d9cf57b36ccf9a675c6b2b070d56c9b297
Remove import of private _lazy module

New version of oslo.i18n released and some things
moved around (internal private modules in the lib).

This should be fine and shouldn't matter to us, BUT
it seems we had some hackery going on in our unit tests
that were being lazy and importing and manipulating the
private library.

This patch removes those cases and fixed up the cinder.i18n
helper method for enable_lazy to accept a bool (which is how
the libraries method works to begin with).

There were several tests in test_faults that were actually
performing and comparing translations of messages.  These
tests aren't quite working now because they had a number
of things they imported from private variables and methods
in the i18n module.  Honestly I'm not sure of the value of
testing those things here anyway, but for now I've just
added a skip to those and we can sort out long term fixes
and plans later.

Change-Id: I2ae3d9b98c107cebaf386adbdcdb3cfafee070be
Partial-Bug: #1408099