Updated openstack/openstack Project: openstack/tempest  6a9764dd73928bc9f6af5718779dcce64540d0df
printout testr tests that fail

We test whether or not testr fails to list all the tests, however
we're not being very nice about reporting when it fails (especially
not what caused the failure), and throw a cryptic 0 != 3 mismatch
error.

This should help make it easy to find syntax issues in the future.

Change-Id: I0d4b5b2a280813d54b0c2712a504e472370bd196