Updated openstack/openstack Project: openstack/oslo.i18n  118aeb0ceb10f80a365db6cd724d9d2720b132f9
Add a fixture to let tests manage the lazy flag

We do not want to expose the stored lazy flag, but we do want to allow
applications to force it to a certain state and then restore it to its
original test for testing. This new fixture provides a way to do that.

Change-Id: Ica05fd8fc7e687da69d61ea84bf1ceae65db93ce