Updated openstack/openstack Project: openstack/oslo.concurrency  590444c1f51d7c6e7cbb179e0900e48adfdf92d7
Add a reader/writer lock

Taskflow has a reader/writer lock that is likely useful
to other projects; and it seems better at home in this
module.

The class provides a way to create reader/writer locks
where there may be many readers at the same time (but
only one writer). It does not allow (currently) for privilege
escalation (but this could be added with limited support
in the future).

Change-Id: Ie763ef92f31c34869d83a533bc8761b0fbd77217