Updated openstack/openstack Project: openstack/keystone  31807b4cd69c9ffe08b416505eedb4729aab746f
Ensure manager grant methods throw exception if role_id is invalid.

We currently don't test whether the assignment manager/driver
grant methods correctly throw a RoleNotFound exception if the role
is invalid. This patch adds a test for this.

Such a test will be useful to ensure this functionality is
maintained as we migrate roles to their own backend.

Change-Id: I51e87129fb3efcc1abae9c59915eaca38eb91227
Closes-bug: 1406721