Updated openstack/openstack Project: openstack/neutron  43e083a1a13de594afa0a77fa74cdc16dd3d72b0
Replace mention of nose with nose2 in devref

Documentation for the nose test runner was previously included in the
testing section of the developer reference.  Due to nose's lack of
support for the load_tests protocol - required to support
testscenarios-based generative testing - the documentation has been
updated to suggest the load_tests-supporting nose2 instead.

Change-Id: I5beb8619d5045372472021ed7eb19bddd0de5862