Updated openstack/openstack Project: openstack/tempest-lib  87c1442469e76e338ee7eb0606e7e8caa525e056
Fix subunit-trace on python < 2.7

This commit fixes running subunit-trace on older versions of python,
by inlining the total_seconds method from newer versions of python.
While tempest-lib is nominally only for python > 2.7 (which is all we
test) subunit-trace is slightly different. Since it is expected to be
used in the unit test workflows of projects, some of which still run
on older python.

Change-Id: I33c337bfb376a27bf02cc2d16a69dafba2cea41f