Updated openstack/openstack Project: openstack/neutron  1210d1c0595ca2888f147ee38257f7f3be96214f
Add Process class helper to manage processes with namespace

This class extends Popen class with usage of namespace and root-helper.
Because of usage of root wrapper, this class re-uses get_child_pid() for
killing the child process. get_child_pid() is taken out of AsyncProcess
as a part of this patch.

Change-Id: I856b3ec75f347ecccaf4a1c6fd17b28a33ee1a3f
Related-Bug: 1243216