Updated openstack/openstack Project: openstack/oslo.db  cfbe5c56340c0e4a3b18902d7e55bb18a45857ec
Ensure mysql_sql_mode is set for MySQLOpportunisticTests

The default setting for oslo.db is TRADITIONAL for mysql_sql_mode;
this is set up in oslo_db/options.py as well as
in EngineFacade; however, the provisioning system currently does not
use these options or EngineFacade, and the default for
mysql_sql_mode in the base session.create_engine() is None.
For a MySQL/MariaDB database where this is not established in my.cnf,
tests will run without TRADITIONAL SQL mode and some tests will fail,
including some Nova migration tests that rely upon this.

This patch resolves that as well as the warning we see
in tests regarding the empty SQL mode.

Change-Id: I3acdc0fb30327f56a76ee299cc6bade7c5a7e312