Updated openstack/openstack Project: openstack/requirements  614bb9a54db2392cb1c2bb026f7df643f7eb1cf8
Angular xstatic for Version 1.3.7

Adding version 1.3.7
1.3.x angular is needed for Horizon launch instance work.

This was dependent on the following (done):
https://review.openstack.org/#/c/142520/

Xstatic-Angular 1.3.7 includes the following packages:

XStatic-Angular-Cookies
XStatic-Angular-Mock

They can be removed, but this apparently needs to be done
in a multi-step process. The interdependency chain,
(horizon has dependency on the old packages) causes
an integration test to fail. So, if we follow the normal
process, then we have to do it in the following order.

1) The newer pakage has to be added to global requirements
2) a) The new dependency has to be added to Horizon
   b) Old dependencies must be removed from Horizon
3) The older packages must be removed from global requirements
   - change 145096

At least this what I can tell how to make it all
work following processes since I can't modify the
requirements across projects in the same patch set.

Partially Implements: blueprint launch-instance-redesign

Change-Id: Ibd171723c6b7ba672c90daf35ae88f672e8d5f4e