Updated openstack/openstack Project: openstack/nova  232cc97a17c2e6e9924c4581ffe50f02cb114240
Add obj_as_admin() to NovaPersistentObject

Since this is a nova-ism, I'm adding this to our subclass of NovaObject,
which we use for almost everything, and at least everything that will care
about this. Later, we should probably generalize a layer between the base
and NovaPersistentObject classes for things like this.

Related to blueprint kilo-objects

Change-Id: I0aed6e19446febd3350b15f63a2855604dd9d5c3