Updated openstack/openstack Project: openstack-infra/devstack-gate  13da1170a0c12cb9af4d36a28a6151478168de6b
Use private addresses everywhere with aiopcpu

For consistency of debugging issues like nova instance resize and
migrations it helps a lot to use a single set of IPs for OpenStack
services. We can't use the public IP for everything (for reasons) but
can use the private addresses everywhere so use them.

This requires setting HOST_IP explicitly in localrc as devstack will
pick the public IPs on nodes that have them by default.

Change-Id: I0c8cf8f5627d1c00f278827ab0cee708f7741e89