Updated openstack/openstack Project: openstack/neutron  6df7f5abf16f11abe784ab33e7885ec7bb8a5273
Deletes floating agent gw port on disassociate

This patch deletes the floating agent gw port
on floatingip disassociate.

A recent change in the neutron trunk broke
this functionality. This was introduced by
the patch that addressed the re-order
operations in (l3_dvr) update floating ip.

The current patch fixes the problem.

Change-Id: I98e766cf4ed33412d9d5dae07217553acbb5009c
Closes-Bug: #1394026