Updated openstack/openstack Project: openstack/nova  68e417f1b17927e2f18e39751987b99ba37b2e78
ec2: Change FormatMappingTestCase to NoDBTestCase

The test doesn't hit the database at all so make it
a NoDBTestCase.

As far as I can tell this is the last ec2 API unit test
class that isn't a trivial convert to NoDBTestCase, all
the rest require some mocking of the objects or DB API
layers to convert them to NoDBTestCase, which I plan to
work on in later patches.

Change-Id: I682a377d769312e8f37f874aa4548665b3b52ed3