Updated openstack/openstack Project: openstack/swift  92fd28aa6afdb97a23027a2f02631eaf693a41b4
Fix a few edges where we lost python 2.6 support

a.k.a. put my head in the sand about the reality of not supporting python 2.6
a little while longer.  We need to get something in the next release notes
about deprecating support for python 2.6 ASAP.  I don't really care enough
about it to keep cleaning up the junk we're going to let slip through not
testing python 2.6 in the gate.

Change-Id: Ib36cd66bda29d75d3b5f4ef0a0ef7b824923df28