Updated openstack/openstack Project: openstack/oslo.concurrency  df35680b673eabe0b664f17663f2272711b2e5b4
add watchdog module

This provides the watchdog.watch function which is a contextmanager
that can be used before doing operations that have a high risk of
deadlocking or aberrantly taking huge amounts of time to complete
(which could be the root cause for other issues).

Unit tests require the use of sleep to make the timers pop either
before or after the event in question. There is 1 second buffer that's
encoded into each scenario which should be sufficient for unit tests
(even in busy systems). If that turns into an issue later, the buffer
can be increased (I was only seeing real skew of 1-2 ms locally).

Change-Id: I79e39f8a4aa702627ca39bea3e6fc672134b8a72