Updated openstack/openstack Project: openstack/keystone  d9bbff295df29f9b09f6a0894cc0acdd6a1e653b
Fix tests using extension drivers

The tests were assuming that extension drivers (such as federation
and revoke_api) were always loaded. These drivers were only
available because a previous test had imported the module and its
factory method is stored in global state in the dependency module.

The base Testcase is changed to not set up the saml2 and oauth1
auth plugins by default. The saml2 auth plugin depends on the
federation API which is only available when the extension is
loaded; similarly for the oauth1 auth plugin.

A new method is added to base TestCase to allow the TestCase to
provide more drivers to load during setUp. The TestCases that need
to load an optional backend make use of it.

Closes-Bug: #1408389
Change-Id: I5774fbe136849b4d1e80e64f3791dc295458e446