Updated openstack/openstack Project: openstack/neutron  eb17ac2902dbee86bf6b37d5fb396b23828590ea
Remove SELECT FOR UPDATE from delete_network and delete_subnet

delete_network and delete_subnet instead of using SELECT FOR UPDATE
delete the ports that are allowed to be auto deleted straight away.
Then a check is performed, if there are ports still associated
with the network or the subnet that cannot be autodeleted,
an exception is raised. The operation will be rolled back.

Partial-bug: #1331564

Change-Id: I29178204b3a220c1c6010f59272981fb1b4099fe