Updated openstack/openstack Project: openstack-dev/devstack  69ef526ac9fbb83c9b6e0cf6d5fc6e65f20fcf54
Enroll Ironic nodes /w IRONIC_KEY_FILE

Do not compose a full path to the ironic ssh keyfile.  Instead,
use IRONIC_KEY_FILE which will be set to the full path to the key.
This is required when IRONIC_KEY_FILE is user-specified, otherwise
we're still enrolling nodes with the default path that is only
used when generating a key.

Change-Id: Icf1ba97becc3e800caf9329a16b79cf106ac3c9a