Updated openstack/openstack Project: openstack/python-swiftclient  7709fea51e71084163cf56c4d11a280ca4388357
Check that content_type header exists before using

When downloading an object, the content_type header is inspected
to determine if a directory needs to be created. This header is
assumed to always be in the response; if it isn't an Exception is
raised. swiftclient should not assume content_type will always be
set.

Change-Id: I156195c02b6e0bc398fa962eb1f78c4dbddd1596