Updated openstack/openstack Project: openstack/horizon  bca3b68bb89818354ae1db97688729738e3ea08c
Fix method passed to selenium way it suggests

A lambda function is passed to selenium wait call until() but the docstrings
of that call suggest a better way of passing the lambda function.
Its redundant to return a specific driver instance when until() function
already takes care of that.

Change-Id: Ie0916587bf6a7488b3828d91ad553a3944619dfc
Partially implements blueprint: selenium-integration-testing
Closes-Bug: #1404416