Updated openstack/openstack Project: openstack/cinder  a45b4036d69186a43927be2c90f435e45c41e2de
Add debug messaging for tgt already exists

We have a race condition where the call to create an
iscsi target seems to be issued twice.  This patch
adds some debug logging to try and get some more info
around where things are going wrong.

First, we add a log message to the intialize_connection call
in the cinder.volume.api and include the connector info.

Second, we add a target show output prior to the update/create
command in the target driver.

In addition, we also had the update and show commands in the same
try/except block which led us to believe the show was failing,
but that's NOT the case at all.  So this patch also moves that
existing show (debug purposes as well) out of the main try block.

Note that there are some old bugs logged against TgtAdm for issues
with it "loosing count" of target id's it's issued, and in turn
reissuing the same ID multiple times.  This patch is intended to
give us enough info to determine if that's what's going on or not.

Other possibility is that the initialize call is being issued more
than once.

Change-Id: If17acff6c0b697261ae0e7a40fc26962cbee6c7b
Related-Bug: #1398078