Updated openstack/openstack Project: openstack/horizon  2f22933b1143f050dd258b4d869bbf77c515923a
Automatically expand the current panel group on dashboard expansion

If the panel is active, then collapse the panelgroup and
collapse the dashboard, the panelgroup with active panel
can not expand while expanding the dashboard which collapsed
just now.

Steps to reproduce:

1. login as admin
2. click 'Project Dashboard',it default expand 'Compute' panel.
3. click 'Network' panel header,then click 'Router'
4. click 'Network' panel header again(collapse the Network panel)
5. click 'Project Dashboard'(collapse the Project panel)
6. click 'Project Dashboard'(expand the Project panel)

After those steps we can see the 'Network' panelgroup can not
expand Automatically, the patch fix it.

Change-Id: I3b4d86c4f7ae42da241d021c746dbe3b8564a4c4
Closes-Bug: 1405681