Updated openstack/openstack Project: openstack/horizon  50009342970146a882af278e6d642d98d4adb801
Add support for volume transfers

Cinder supports volume transfers, which are transfers of ownership from
one project to another.  The normal flow is that a donor creates a volume
transfer, captures the resulting transfer id and secret authentication
key, and passes that information out of band to the recipient (such as
by email, text message, etc.).  The recipient then accepts the transfer,
supplying the transfer id and authentication key.  The ownership of the
volume is then transferred from the donor to the recipient, and the
volome is no longer visible to the donor.

It is important to know of some limitations about the cinder API for
volume transfers and their impact on the UI design.
1. When creating a volume transfer, you cannot specify who the intended
   recipient will be, and anyone with the transfer ID and auth key can
   claim the volume.  Therefore there is no UI to prompt for a
   recipient.
2. Current volume transfers are only visible to the donor; users in
   other projects are unable to view these transfers.  This makes it
   pointless to create a project table to view and accept volume
   transfers, since the current transfers are not visible.  Instead, the
   transfer information has been added to the volume details, which are
   visible by the donor, and the volume state clearly reflects that a
   transfer has been created.  It it also impossible to present to the
   recipient a pull-down list of transfers to accept.
3. The only time that the authorization key is visible to the donor is
   in the response from the creation of the transfer; after creation it
   is impossible for even the donor to recover it.  Since the donor must
   capture the transfer ID and authorization key in order to send it to
   the recipient, an extra form was created to present this information
   to the donor immediately after the transfer has been created.  It is
   not enough to put the information in a success message, since these
   are typically only displayed for a small number of seconds before
   they disappear.

Co-Authored-By: Gary W. Smith <gary.w.smith@hp.com>
Co-Authored-By: Julie Gravel <julie.gravel@hp.com>

Implements blueprint cinder-volume-transfer

Change-Id: I0134f90c93a8d804d2c87fe9713bcef9e1629d79