Updated openstack/openstack Project: openstack/nova  ad51b22c74114f46e0758b0a074723f6e68c4b74
Add LibvirtGPFSVolumeDriver class

Currently, GPFS use LibvirtVolumeDriver to generate the libvirt
xml configuration when attaching a gpfs volume. But the driver
always set the disk_type to 'block' which should be 'file' for
GPFS volume.

This patch create a new volume driver class for GPFS, the class
can generate the right xml configuration for GPFS volume.

Change-Id: Ica1cf3558baeea12b519bda50c61cea446429364
Closes-Bug:#1405044